Modified project structure for a succesful merge
